+++ src/main/java/com/communityratesgames/rating/RatingRepository.java
@@ -18,5 +18,5 @@ public interface RatingRepository extends JpaRepository<RatingEntity, Long> {
     RatingEntity findByGameAndUser(GameEntity game, UserEntity user);
 
     @Query(value = "SELECT AVG(r.rating) FROM RatingEntity r WHERE r.game = :gameId")
-    long getGameAverageRating(@Param("gameId") long gameId);
+    Float getGameAverageRating(@Param("gameId") long gameId);
 }
